Refactor customer types into seperate file
